requirements/tox_requirements27.txt
mock==3.0.5
requests_oauthlib==1.2.0
cryptography==2.8
ddt==1.2.2
django==1.11.24
django-oauth-toolkit==1.0.0  # not needed in production, unless you need to create provider's Applications
psycopg2==2.8.3
factory_boy==2.12.0
pycodestyle==2.5.0
pylint==1.7.6
edx_lint==1.3.0
diff-cover==2.3.0
coverage==4.5.4
requests-mock==1.7.0
testfixtures==6.10.0
